Move export-utils to ui/helpers as it's only used in the UI now.
