+++ ui/components/ui/export-text-container/export-text-container.component.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Copy from '../icon/copy-icon.component';
 import { useI18nContext } from '../../../hooks/useI18nContext';
 import { useCopyToClipboard } from '../../../hooks/useCopyToClipboard';
-import { exportAsFile } from '../../../../shared/modules/export-utils';
+import { exportAsFile } from '../../../helpers/utils/export-utils';
 
 function ExportTextContainer({
   text = '',
